[Drone_Simulation] Added handling of alerting messages

+++ src/main/scala/utils/Message.scala
@@ -18,7 +18,7 @@ Registration State
 case class DroneMsg (
     Issue_Date: String, // maybe more advanced Date java/scala type
     Plate_ID: String, // GGY6450 is a license plate
-    Violation_Code: Byte, // [0,99]
+    Violation_Code: Int, // [-1,99]
     Vehicle_Body_Type: String,
     Street_Code1: Int, // [0,98K]
     Street_Code2: Int, // [0,98K]
